requirements.txt
torch
torchvision
flask
requests
PILLOW
